Cleyson

+++ application/views/cadastros/forma-lista.php
@@ -21,46 +21,60 @@
                     <tr>
                         <th class="tabela_header">Nome</th>
                         <th class="tabela_header" width="70px;" colspan="2"><center>Detalhes</center></th>
-                    </tr>
+                </tr>
                 </thead>
                 <?php
-                    $url      = $this->utilitario->build_query_params(current_url(), $_GET);
-                    $consulta = $this->forma->listar($_GET);
-                    $total    = $consulta->count_all_results();
-                    $limit    = 10;
-                    isset ($_GET['per_page']) ? $pagina = $_GET['per_page'] : $pagina = 0;
+                $url = $this->utilitario->build_query_params(current_url(), $_GET);
+                $consulta = $this->forma->listar($_GET);
+                $total = $consulta->count_all_results();
+                $limit = 10;
+                isset($_GET['per_page']) ? $pagina = $_GET['per_page'] : $pagina = 0;
 
-                    if ($total > 0) {
-                ?>
-                <tbody>
-                    <?php
+                if ($total > 0) {
+                    ?>
+                    <tbody>
+                        <?php
                         $lista = $this->forma->listar($_GET)->orderby('descricao')->limit($limit, $pagina)->get()->result();
                         $estilo_linha = "tabela_content01";
                         foreach ($lista as $item) {
                             ($estilo_linha == "tabela_content01") ? $estilo_linha = "tabela_content02" : $estilo_linha = "tabela_content01";
-                     ?>
+                            ?>
                             <tr>
                                 <td class="<?php echo $estilo_linha; ?>"><?= $item->descricao; ?></td>
 
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
-                                    <a href="<?= base_url() ?>cadastros/forma/carregarforma/<?= $item->forma_entradas_saida_id ?>">Editar</a>
-                            </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
-                                    <a onclick="javascript: return confirm('Deseja realmente exlcuir esse Forma?');" href="<?= base_url() ?>cadastros/forma/excluir/<?= $item->forma_entradas_saida_id ?>">Excluir</a>
-                            </td>
-                        </tr>
+                                <?
+                                $perfil_id = $this->session->userdata('perfil_id');
+                                ?>
+                                <? if ($perfil_id != 10) { ?>
+                                    <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                        <a href="<?= base_url() ?>cadastros/forma/carregarforma/<?= $item->forma_entradas_saida_id ?>">Editar</a>
+                                    </td>
+                                    <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                        <a onclick="javascript: return confirm('Deseja realmente exlcuir esse Forma?');" href="<?= base_url() ?>cadastros/forma/excluir/<?= $item->forma_entradas_saida_id ?>">Excluir</a>
+                                    </td>
+                                    
+                                    <?}else{?>
+                                    <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                        Editar
+                                    </td>
+                                    <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                        Excluir
+                                    </td>
+                                    
+                                <?}?>
+                                </tr>
 
-                        </tbody>
-                        <?php
-                                }
-                            }
-                        ?>
-                        <tfoot>
-                            <tr>
-                                <th class="tabela_footer" colspan="4">
-                                   <?php $this->utilitario->paginacao($url, $total, $pagina, $limit); ?>
-                            Total de registros: <?php echo $total; ?>
-                                </th>
+                            </tbody>
+                            <?php
+                        }
+                    }
+                    ?>
+                    <tfoot>
+                        <tr>
+                            <th class="tabela_footer" colspan="4">
+                                <?php $this->utilitario->paginacao($url, $total, $pagina, $limit); ?>
+                                Total de registros: <?php echo $total; ?>
+                        </th>
                     </tr>
                 </tfoot>
             </table>
@@ -70,8 +84,8 @@
 </div> <!-- Final da DIV content -->
 <script type="text/javascript">
 
-    $(function() {
-        $( "#accordion" ).accordion();
+    $(function () {
+        $("#accordion").accordion();
     });
 
 </script>
